[#231] Refactor: Clean up all index.ts files from k8s
Cleaned up index.ts files from src/k8s folder
