+++ src/brokers/broker-details/components/BrokerDetailsBreadcrumb/BrokerDetailsBreadcrumb.tsx
@@ -12,7 +12,7 @@ import {
 } from '@patternfly/react-core';
 import { useTranslation } from '../../../../i18n/i18n';
 import { k8sDelete } from '@openshift-console/dynamic-plugin-sdk';
-import { AMQBrokerModel } from '../../../../k8s';
+import { AMQBrokerModel } from '../../../../k8s/models';
 import { PreConfirmDeleteModal } from '../../../view-brokers/components/PreConfirmDeleteModal/PreConfirmDeleteModal';
 import { useNavigate } from 'react-router-dom-v5-compat';
