+++ src/brokers/broker-pods/components/PodList.tsx
@@ -8,7 +8,7 @@ import {
   TableColumn,
 } from '@openshift-console/dynamic-plugin-sdk';
 import { useTranslation } from '../../../i18n/i18n';
-import { BrokerCR } from '../../../k8s';
+import { BrokerCR } from '../../../k8s/types';
 import { PodRow } from './PodRow';
 
 type PodsTableProps = {
