+++ src/brokers/update-broker/UpdateBroker.container.tsx
@@ -3,7 +3,8 @@ import { k8sGet, k8sUpdate } from '@openshift-console/dynamic-plugin-sdk';
 import { AlertVariant } from '@patternfly/react-core';
 import { AddBroker } from '../add-broker/AddBroker.component';
 import { Loading } from '../../shared-components/Loading/Loading';
-import { AMQBrokerModel, BrokerCR } from '../../k8s';
+import { AMQBrokerModel } from '../../k8s/models';
+import { BrokerCR } from '../../k8s/types';
 import {
   ArtemisReducerOperations,
   BrokerCreationFormDispatch,
