+++ src/brokers/view-brokers/components/BrokersList/BrokersList.tsx
@@ -10,7 +10,7 @@ import {
 } from '@openshift-console/dynamic-plugin-sdk';
 import { BrokerRow, BrokerRowProps } from './BrokerRow';
 import { useTranslation } from '../../../../i18n/i18n';
-import { BrokerCR } from '../../../../k8s';
+import { BrokerCR } from '../../../../k8s/types';
 
 type BrokersTableProps = Pick<
   BrokerRowProps,
