+++ src/reducers/7.12/reducer.ts
@@ -4,7 +4,7 @@ import {
   BrokerCR,
   Acceptor,
   ResourceTemplate,
-} from '../../k8s';
+} from '../../k8s/types';
 import { createContext } from 'react';
 import { SelectOptionObject } from '@patternfly/react-core';
 import { ConfigType } from '../../shared-components/FormView/BrokerProperties/ConfigurationPage/ConfigurationPage';
