+++ src/shared-components/FormView/BrokerProperties/ConfigurationPage/AcceptorsConfigPage/AcceptorConfigSection/AcceptorConfigPage/ListPresets/ListPresets.tsx
@@ -5,7 +5,7 @@ import {
 } from '@patternfly/react-core';
 import { FC, useContext } from 'react';
 import { useTranslation } from '../../../../../../../../i18n/i18n';
-import { Acceptor } from '../../../../../../../../k8s';
+import { Acceptor } from '../../../../../../../../k8s/types';
 import {
   ArtemisReducerOperations,
   BrokerCreationFormDispatch,
@@ -13,7 +13,7 @@ import {
   getAcceptorFromCertManagerResourceTemplate,
   getCertManagerResourceTemplateFromAcceptor,
 } from '../../../../../../../../reducers/7.12/reducer';
-import { ResourceTemplate } from '../../../../../../../../k8s';
+import { ResourceTemplate } from '../../../../../../../../k8s/types';
 import { SelectIssuerDrawer } from '../../SelectIssuerDrawer/SelectIssuerDrawer';
 import { ConfirmDeleteModal } from '../../ConfirmDeleteModal/ConfirmDeleteModal';
 type ResourceTemplateProps = {
