+++ src/shared-components/FormView/BrokerProperties/ConfigurationPage/AcceptorsConfigPage/AcceptorConfigSection/PresetButton/PresetButton.tsx
@@ -16,7 +16,7 @@ import {
 import { CSSProperties, FC, useContext, useState } from 'react';
 
 import { useTranslation } from '../../../../../../../i18n/i18n';
-import { Acceptor } from '../../../../../../../k8s';
+import { Acceptor } from '../../../../../../../k8s/types';
 import {
   ArtemisReducerOperations,
   BrokerCreationFormDispatch,
