+++ src/shared-components/FormView/BrokerProperties/ConfigurationPage/AcceptorsConfigPage/AcceptorConfigSection/SelectIssuerDrawer/SelectIssuerDrawer.tsx
@@ -1,10 +1,7 @@
 import { useTranslation } from '../../../../../../../i18n/i18n';
 import { BrokerCreationFormState } from '../../../../../../../reducers/7.12/reducer';
-import {
-  CertIssuerModel,
-  CertModel,
-  IssuerResource,
-} from '../../../../../../../k8s';
+import { CertIssuerModel, CertModel } from '../../../../../../../k8s/models';
+import { IssuerResource } from '../../../../../../../k8s/types';
 import {
   RedExclamationCircleIcon,
   k8sCreate,
